test(omo-codex): split lazycodex install surface coverage
